The string ?url= should only be used with PROXY_SCRIPT
The new standard for Forecast.io is to use a non query parameter url,
therefore query strings were removed. The structure for the proxy
script can still be created, use what is passed, but the building of
the URL needed to be moved before the insertion of Geo coordinates.

Also, getTime() is working, I’m unsure why it was disabled.